fix: address ChatGPT UI review SUGGESTIONs (CI, docs, test dedup)
SUGGESTIONs from the ChatGPT UI review (non-blocking). All
follow-ups from the BLOCKER fix commit.

#1 CI: smoke-test job was unreachable
  - Added push.tags: [v*] and workflow_dispatch: triggers so
    the smoke-test job can actually run.
  - Updated the 'if' to accept either: tag pushes matching v*
    or workflow_dispatch events.

#2 README: saliva + OpenMMConfig construction referenced
   removed cacl2_mol / kh2po4_mol fields
  - Reworded the saliva preset description to NaCl-only.
  - Replaced the 'nacl_mol, cacl2_mol, kh2po4_mol, ...' list with
    'nacl_mol, protonation_ph, temperature_k'.
  - Added an Ionic Strength caveat explaining Ca2+ / KH2PO4 are
    unmodelled context, not supported config.

#3 AGENTS.md: stale architecture + coverage floor
  - Removed 'EquilibrationStage' from the openmm/config.py
    description.
  - Added 'paths.py # Centralized filenames (FileNames)' to the
    module map.
  - Updated quick-reference coverage entry 55% -> 70%.

#4 CHANGELOG.md: stale details
  - Added date '2026-07-25' to the [Unreleased] entry.
  - Corrected coverage floor (55% -> 70%), CI command
    (--all-extras -> --extra openmm --group dev for unit),
    RUF001/002/003 reason (Unicode = domain notation).
  - Added 'Removed dead config' + 'CHARMM detection' +
    'resolve_pdb promotion' entries for the Phase 8 work.

#6 test dedup: TestPbcCorrectTriclinic was a near-duplicate
   of test_geometry.py
  - Moved the 4 PBC tests (orthorhombic parity,
    dodecahedron-diagonal, dodecahedron-face-crossing,
    broadcasting) to test_geometry.py with the pre-fix
    regression assertions preserved.
  - Replaced test_openmm_runner.py::TestPbcCorrectTriclinic
    with a TestPbcCorrectMoved marker class so a search for
    'TestPbcCorrect' in test_openmm_runner.py still finds a
    pointer to the canonical location.
  - Removed the now-unused pbc_correct + dodecahedron_box
    imports from test_openmm_runner.py.

#7 CHARMM detection test
  - Added test_non_charmm_prefix_not_misclassified: a
    protein_ff='non-charmm-test' must NOT take the CHARMM
    XML branch. Pins the new prefix-match behavior so the
    substring-vs-prefix fix can't regress.

All 'make validate' checks pass. 167 tests.
