+++ .github/workflows/ci.yml
@@ -3,8 +3,10 @@ name: CI
 on:
   push:
     branches: [main]
+    tags: ["v*"]
   pull_request:
     branches: [main]
+  workflow_dispatch:
 
 permissions:
   contents: read
@@ -52,9 +54,16 @@ jobs:
   # runners). Trigger manually with `workflow_dispatch` after the lint
   # gate is green. Run on tagged releases for an end-to-end check.
   smoke-test:
+    # Triggered by tag pushes (v*) or manual dispatch — NOT on regular
+    # PRs or main-branch pushes. Requires a self-hosted runner with a
+    # CUDA-capable GPU + OpenMM built with CUDA support (not available
+    # on GitHub-hosted runners). Run ``workflow_dispatch`` from the
+    # Actions tab to invoke manually.
     runs-on: [self-hosted, gpu]
     needs: lint-and-test
-    if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags/')
+    if: |
+      (github.event_name == 'push' && startsWith(github.ref, 'refs/tags/v')) ||
+      github.event_name == 'workflow_dispatch'
     steps:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6
       - name: Install uv
